for to stream

+++ 4/src/test/java/com/twu/refactoring/NumberCruncherTest.java
@@ -1,6 +1,7 @@
 package com.twu.refactoring;
 
-import org.junit.jupiter.api.Test;
+
+import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -8,25 +9,25 @@ import static org.hamcrest.MatcherAssert.assertThat;
 public class NumberCruncherTest {
     @Test
     public void shouldCountEvenNumbers() {
-        int evens = new NumberCruncher(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11).countEven();
-        assertThat(evens, is(5));
+        long evens = new NumberCruncher(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11).countEven();
+        assertThat(evens, is(5L));
     }
 
     @Test
     public void shouldCountOddNumbers() {
-        int odds = new NumberCruncher(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11).countOdd();
-        assertThat(odds, is(6));
+        long odds = new NumberCruncher(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11).countOdd();
+        assertThat(odds, is(6L));
     }
 
     @Test
     public void shouldCountPositiveNumbers() {
-        int positives = new NumberCruncher(-5, -4, -3, -2, -1, 0, 1, 2, 3, 4).countPositive();
-        assertThat(positives, is(5));
+        long positives = new NumberCruncher(-5, -4, -3, -2, -1, 0, 1, 2, 3, 4).countPositive();
+        assertThat(positives, is(5L));
     }
 
     @Test
     public void shouldCountNegativeNumbers() {
-        int negatives = new NumberCruncher(-5, -4, -3, -2, -1, 0, 1, 2, 3, 4).countNegative();
-        assertThat(negatives, is(5));
+        long negatives = new NumberCruncher(-5, -4, -3, -2, -1, 0, 1, 2, 3, 4).countNegative();
+        assertThat(negatives, is(5L));
     }
 }
